Правки в код-стайл
