+++ src/httphandler/BaseHttpHandler.java
@@ -1,4 +1,4 @@
-package http_handler;
+package httphandler;
 
 import com.google.gson.Gson;
 import com.sun.net.httpserver.HttpExchange;
