+++ src/httphandler/PrioritizedHttpHandler.java
@@ -1,4 +1,4 @@
-package http_handler;
+package httphandler;
 
 import com.google.gson.Gson;
 import com.sun.net.httpserver.HttpExchange;
@@ -7,13 +7,13 @@ import manager.TaskManager;
 
 import java.io.IOException;
 
-public class HistoryHttpHandler extends BaseHttpHandler implements HttpHandler {
-    public HistoryHttpHandler(TaskManager taskManager, Gson gson) {
+public class PrioritizedHttpHandler extends BaseHttpHandler implements HttpHandler {
+    public PrioritizedHttpHandler(TaskManager taskManager, Gson gson) {
         super(taskManager, gson);
     }
 
     @Override
     public void handle(HttpExchange exchange) throws IOException {
-        writeResponse(exchange, gson.toJson(taskManager.getHistory()), 200);
+        writeResponse(exchange, gson.toJson(taskManager.getPrioritizedTasks()), 200);
     }
 }
